refactor: Rename `native_coin` -> `native_currency`
Avoid confusion with `neptune_coins`, which is for an amount of the
native currency.
